was moved components
