--- src/common/components/AssetTransaction/components/index.ts
@@ -1,1 +1,0 @@
-export { default as TableHeader } from './TableHeader/TableHeader';
