+++ src/common/components/TableHeader/TableHeader.module.scss
